chore: rename connect-wallet-styles to styles
